refactor: remove unused sources
