+++ data/src/main/java/vn/root/data/di/LocalModule.kt
@@ -1,4 +1,4 @@
-package vn.root.data.di.local
+package vn.root.data.di
 
 import android.content.Context
 import dagger.Module
@@ -10,6 +10,7 @@ import vn.core.data.local.CoreDatabase
 import vn.core.data.local.PreferenceWrapper
 import vn.root.data.BuildConfig
 import vn.root.data.local.AppDatabase
+import vn.root.data.local.ItemDao
 import javax.inject.Singleton
 
 @Module
@@ -25,4 +26,8 @@ internal class LocalModule {
     @Singleton
     fun provideAppDatabase(@ApplicationContext context: Context): AppDatabase =
         CoreDatabase.build<AppDatabase>(context, "root-database.db")
+
+    @Provides
+    @Singleton
+    fun provideItemDao(appDatabase: AppDatabase): ItemDao = appDatabase.itemDao()
 }
