--- data/src/main/java/vn/root/data/di/qualifier/AnoRetrofitApiService.kt
@@ -1,8 +1,0 @@
-package vn.root.data.di.qualifier
-
-import javax.inject.Qualifier
-
-@Qualifier
-@MustBeDocumented
-@Retention(AnnotationRetention.RUNTIME)
-annotation class AnoRetrofitApiService
